+++ libs/agentos-api-client/src/lib/model/agent-config-search-request.ts
@@ -8,9 +8,7 @@
  * Do not edit the class manually.
  */
 
-export interface UserGroupUpdateRequest {
-  name: string
-  userExternalIdsToAdd: Set<string>
-  userExternalIdsToRemove: Set<string>
-  agentIds: Set<string>
+export interface AgentConfigSearchRequest {
+  namespaceExternalId: string
+  userExternalId: string
 }